refactor: change structure, refactor methods
