+++ commands/exit.js
@@ -1,6 +1,6 @@
 import { rl } from "../index.js";
 
-export const exitCommand = (args) => {
-  console.log(`Thank you for using File Manager, ${args}, goodbye!`);
+export const exitCommand = (userName) => {
+  console.log(`Thank you for using File Manager, ${userName}, goodbye!`);
   rl.close();
 };
